GDC-Level-8-Milestone-master/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-filter==21.1
redis==4.1.4
celery==4.4.7